Re #1152 other implementation of long_idx.m

+++ herbert_core/utilities/classes/@fast_map/private/optimize_.m
@@ -6,14 +6,14 @@ function obj = optimize_(obj,minmax_keys)
 
 
 obj.min_max_key_ = minmax_keys;
-obj.key_shif_        = obj.min_max_key_(1)-1;
+obj.key_shif_        = double(obj.min_max_key_(1))-1;
 
 n_places = obj.min_max_key_(2)-obj.min_max_key_(1)+1;
 
 val_optimized = nan(1,n_places);
 
 if ~isempty(obj.values_)
-    keys_shifted = obj.keys_-obj.min_max_key_(1)+1;    
+    keys_shifted = obj.keys_-obj.min_max_key_(1)+1;
     val = obj.values_;
     val_optimized(keys_shifted) = val(:);
 end
